// road-assistant-mobile/service/Api/models/CommentResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Coordinates } from './Coordinates';
export type CommentResponse = {
    id: number;
    createdAt: string;
    text: string;
    routeId: number;
    userId: string;
    coordinates: Coordinates;
};